+++ src/main/java/com/project/myacademy/domain/employee/dto/FindPasswordEmployeeRequest.java
@@ -11,6 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class FindPasswordEmployeeRequest {
     private String name;
-    private String account;
     private String email;
 }
