+++ src/main/java/com/project/myacademy/global/exception/ErrorCode.java
@@ -28,6 +28,8 @@ public enum ErrorCode {
 
     ACCOUNT_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 계정명을 찾을 수 없습니다."),
     EMPLOYEE_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 직원을 찾을 수 없습니다."),
+    NAME_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 이름의 직원을 찾을 수 없습니다."),
+    EMAIL_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 이메일의 직원을 찾을 수 없습니다."),
     REQUEST_EMPLOYEE_NOT_FOUND(HttpStatus.NOT_FOUND, "요청한 직원을 해당 학원에서 찾을 수 없습니다."),
     ACADEMY_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 학원을 찾을 수 없습니다."),
     PARENT_NOT_FOUND(HttpStatus.NOT_FOUND, "해당 부모를 찾을 수 없습니다."),
